core_backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
